+++ changelog.md
@@ -15,10 +15,13 @@
 - Optimized the settings objects for future independence from this plugin
 - Usage of external embedded YouTube URLs in Block Editor is now possible
 - Choose your YouTube videos from local media library in Elementors Video widget
+- Remove some unused codes
 
 ### Fixed
 
 - Fixed refreshing of permalinks (which as been run on every request until now)
+- Fixed potential error on WP CLI function to add Google Drive connection
+- Fixed missing support for plugin PreventDirectAccess
 - Fixed some typos
 
 ## [3.0.0] - 15.02.2025